# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sarkari Exams API",
    "version": "1.0.0",
    "description": "Shared API contract for frontend/backend"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/api/auth/csrf": {
      "get": {
        "summary": "GET /api/auth/csrf",
        "tags": [
          "auth"
        ],
        "operationId": "getApiAuthCsrf",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Auth response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "POST /api/auth/logout",
        "tags": [
          "auth"
        ],
        "operationId": "postApiAuthLogout",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "GET /api/auth/me",
        "tags": [
          "auth"
        ],
        "operationId": "getApiAuthMe",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "summary": "Register",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Auth response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/bookmarks": {
      "get": {
        "summary": "List bookmarks",
        "responses": {
          "200": {
            "description": "Bookmarks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add bookmark",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookmarkCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/bookmarks/{id}": {
      "delete": {
        "summary": "Remove bookmark",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Removed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ContentType": {
        "type": "string",
        "enum": [
          "job",
          "result",
          "admit-card",
          "syllabus",
          "answer-key",
          "admission"
        ]
      },
...
      "Tag": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "slug"
        ]
      },
      "ImportantDate": {
        "type": "object",
        "properties": {
          "eventName": {
            "type": "string"
          },
          "eventDate": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "eventName",
          "eventDate"
        ]
      },
      "Announcement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/ContentType"
          },
          "category": {
            "type": "string"
          },
          "organization": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "externalLink": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "deadline": {
            "type": "string",
            "format": "date-time"
          },
          "minQualification": {
            "type": "string"
          },
          "ageLimit": {
            "type": "string"
          },
          "applicationFee": {
            "type": "string"
          },
          "salaryMin": {
            "type": "number"
          },
          "salaryMax": {
            "type": "number"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ]
          },
          "cutoffMarks": {
            "type": "string"
          },
          "totalPosts": {
            "type": "integer"
          },
          "postedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          },
          "viewCount": {
            "type": "integer"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "importantDates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportantDate"
            }
          },
          "jobDetails": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "id",
          "title",
          "slug",
          "type",
          "category",
          "organization",
          "postedAt",
          "isActive",
          "viewCount"
        ]
      },
...
      "AnnouncementListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Announcement"
            }
          },
          "count": {
            "type": "integer"
          }
        },
        "required": [
          "data"
        ]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "createdAt",
          "updatedAt"
        ]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/User"
              }
            },
            "required": [
              "user"
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          },
          "challengeToken": {
            "type": "string"
          },
          "twoFactorCode": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "BookmarkCreateRequest": {
        "type": "object",
        "properties": {
          "announcementId": {
            "type": "string"
          }
        },
        "required": [
          "announcementId"
        ]
      },
...
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  }
}